First Attempt at harder DynamicProblems
Coding in java. I actually quite like it. Despite the aversion I've had to java due to viewing it as a more complex abstacted version of C, I find coding problems in it quite nice, as it lacks the distracting levels of abstraction of Python, but maintains enough functionality to not have to delve into several hour excursions (albeit enjoyable) into accomplishing simple tasks as in C. Very Good.