University Java OOP Practice Exercises
Java's quite nice when you get the hang of it - I'm starting to see why people like it. While occasionally the strict enforcement of objects can be restricting, you get used to it after a while. What was most noticeable was the significant increase in time that it took me to perform certain tasks, that otherwise would have taken me ages in C.
C still remains my favourite language, but for now, this is okay.